!status can now set a list of status, replacing the original ones
