+++ lang/alexis/en.yml
@@ -26,3 +26,4 @@ command-delayed: You still can't use this command
 nsfw-only: This command can only be used on a NSFW channel.
 error-debug: The command raised an exception while executing it
 error-msg: Could not execute this command
+invite-text: 'You can add me to your guild with the following link:'
