+++ modules/config/bot_status.py
@@ -2,20 +2,21 @@ from datetime import datetime
 
 from discord import Game
 
-from bot import Command, AlexisBot
+from bot import Command, AlexisBot, categories
 from bot.utils import deltatime_to_str_short
 
 
 class BotStatus(Command):
     def __init__(self, bot):
         super().__init__(bot)
+        self.name = 'status'
+        self.help = '$[config-status-help]'
+        self.bot_owner_only = True
+        self.category = categories.SETTINGS
+        self.last_status = ''
+
         self.count = 0
-
-    async def on_ready(self):
-        self.bot.schedule(self.task_other, 30)
-
-    async def task_other(self):
-        status_list = [
+        self.status_list = [
             lambda: 'discord.cl/bot',
             lambda: 'version {}'.format(AlexisBot.__version__),
             lambda: 'add with !invite',
@@ -28,10 +29,34 @@ class BotStatus(Command):
             )
         ]
 
-        status = status_list[self.count]()
+        self.custom_list = []
+
+    async def handle(self, cmd):
+        self.custom_list = [] if cmd.argc < 1 else [f.strip() for f in cmd.text.split('|') if f.strip() != '']
+        self.count = 0
+
+        await self.task_other()
+        await cmd.answer('$[config-status-ok]')
+
+    async def on_ready(self):
+        self.bot.schedule(self.task_other, 30)
+
+    async def task_other(self):
+        status = self.pop()
+        if status == self.last_status:
+            return
+
+        self.last_status = status
         self.log.debug('Changing status to "%s"', status)
         await self.bot.change_presence(game=Game(name=status))
 
-        self.count += 1
-        if self.count >= len(status_list):
+    def pop(self):
+        curr_list = self.status_list if len(self.custom_list) == 0 else self.custom_list
+
+        if self.count < 0 or self.count > (len(curr_list) - 1):
             self.count = 0
+
+        item = curr_list[self.count]
+        self.count += 1
+
+        return item if isinstance(item, str) else item()
